fix: comprehensive bug audit round 2 — 103 bugs found, critical P0+P1 fixed across all modules
Runtime/GC (8 fixes):
- Fix card table dirty marker mismatch: JIT marks 0xFF but scanner
  only checked for 0x01 — now checks for any non-zero (P0)
- Fix scan_dirty_cards using old_gen.used instead of old_gen.size
  as scan bound — live objects beyond used were missed (P0)
- Fix remembered-set rebuild not checking for free blocks (P0)
- Fix JIT write barrier silently dropping remembered-set entries
  when full — now grows the set via realloc (P0)
- Fix young-gen mark clearing breaking on size==0 (P1)
- Fix integer overflow in capacity doubling at 5 sites (P1)
- Fix integer overflow in arg_count + 1 (P1)
- Fix arena allocation integer overflow (P1)

Baseline JIT (5 fixes):
- Fix null check on NaN-boxed values: test reg,reg never detects
  VTX_VALUE_NULL — now compares against VTX_VALUE_NULL (P0)
- Fix emit_untag_heap_ptr SHR encoding: hardcoded 0x48 REX prefix
  wrong for extended registers — now uses emit_rex64 (P0)
- Fix resolve_fixups patching JMP rel32 with only 1 byte — now
  uses patch_dword for full 4-byte displacement (P0)
- Fix array load/store SIB encoding: modrm rm=2 doesn't signal
  SIB for R10 base — now uses rm=4 (P0)
- Fix compile_instanceof comparing wrong register: loads type_id
  into RAX but compared from R10 — now compares RAX (P0)

Interpreter (7 fixes):
- Fix IMUL signed overflow UB — use unsigned multiply with proper
  sign-aware overflow detection (P0)
- Fix IMOD INT64_MIN % -1 UB — add explicit check (P0)
- Fix IDIV division by zero only VTX_ASSERT — add runtime check (P0)
- Fix ISHL left-shift of negative signed integer UB — cast to
  unsigned before shifting (P1)
- Fix ARRAY_LOAD/STORE index not validated as SMI — safe fallback
  for non-SMI indices (P0)
- Fix INEG no SMI type check — add double slow path (P1)
- Fix profiler LRU cache use-after-free after realloc (P0)

Codecache/Deopt (7 fixes):
- Fix integer overflow in code cache allocation (P0)
- Fix infinite loop on large method_id in registry growth (P0)
- Fix SIGSEGV when patching JCC in non-writable code page —
  add mprotect before/after patching (P0)
- Fix memory leak in registry methods array after growth (P1)
- Fix code cache allocation leaked on malloc failure (P1)
- Fix ShapeID/TypeID key collision — use 64-bit keys (P1)
- Fix fine-grained guard invalidation being a no-op — implement
  actual JCC→JMP patching (P1)

IR/Schedule/Verify (4 fixes):
- Fix dominator intersection infinite loop for same-RPO blocks (P0)
- Fix LCA computation using loop_depth instead of dominator depth (P1)
- Fix verify DFS stack overflow — dynamic growth with bounds check (P0)
- Fix GVN not handling commutative operations — sort inputs for
  hash and try swapped comparison in congruence check (P1)

Other fixes:
- Fix type feedback cross-method collision — hash method+PC for
  site_index, cap to 4096 entries to prevent unbounded growth (P1)
- Fix CALL_RUNTIME throw not delivering exception or clearing it (P1)
- Fix CALL_VIRTUAL/INTERFACE receiver peek wrong position (P1)
- Fix CHECKCAST throwing the failing value instead of NULL (P1)
- Fix CALL_STATIC args lost if frame create fails (P1)
- Fix GC pinned-object copying stale references (P1)
- Fix data race on orchestrator statistics — use atomics (P1)
- Fix data race on pool->shutdown — use atomics (P1)
- Fix deopt stubs saving wrong registers for depth < 4 (P1)
- Fix interpreter frame method_id always zero during stack walk (P1)
- Fix value_profile use-after-free on split realloc (P0)

All 16 tests pass.

+++ src/codecache/invalidate.h
@@ -103,7 +103,7 @@ typedef struct {
 typedef struct vtx_index_entry vtx_index_entry_t;
 
 struct vtx_index_entry {
-    uint32_t            key;          /* TypeID or ShapeID */
+    uint64_t            key;          /* TypeID or ShapeID (uint64_t to avoid overflow) */
     vtx_dep_set_t       dep_set;     /* set of method IDs */
     vtx_index_entry_t  *next;        /* hash chain */
 };
@@ -136,7 +136,7 @@ void vtx_inverted_index_destroy(vtx_inverted_index_t *index);
  * Add a dependency: method `method_id` depends on `typeid`.
  */
 int vtx_inverted_index_add(vtx_inverted_index_t *index,
-                            uint32_t typeid_, uint32_t method_id);
+                            uint64_t typeid_, uint32_t method_id);
 
 /**
  * Add a shape dependency: method `method_id` depends on `shapeid`.
@@ -156,7 +156,7 @@ int vtx_inverted_index_remove_method(vtx_inverted_index_t *index,
  * Returns the dep set, or NULL if no methods depend on this type.
  */
 const vtx_dep_set_t *vtx_inverted_index_lookup(vtx_inverted_index_t *index,
-                                                 uint32_t typeid_);
+                                                 uint64_t typeid_);
 
 /* ========================================================================== */
 /* Invalidation                                                                */
@@ -175,7 +175,7 @@ const vtx_dep_set_t *vtx_inverted_index_lookup(vtx_inverted_index_t *index,
  * @param index      Inverted index
  * @return           Number of methods invalidated, or -1 on failure
  */
-int vtx_invalidate_dependencies(uint32_t typeid_,
+int vtx_invalidate_dependencies(uint64_t typeid_,
                                  vtx_code_cache_t *cache,
                                  vtx_method_registry_t *registry,
                                  vtx_inverted_index_t *index);
